\name{E}
\alias{E}
\title{Calcola l'indice di eterogeneita'}
\description{
  Questa funzione permette il calcolo dell'indice di eterogeneita' di Gini.
}
\usage{
E(x)
}
\arguments{
  \item{x}{vettore di dati}
}
\seealso{
  \code{\link{var}}.
}
\examples{
x <- c("A", "A", "B", "A", "C", "A")
E(x)
}
\keyword{univar}
