\name{mean.a}
\alias{mean.a}
\title{Calcola la media armonica}
\description{
  Questa funzione permette il calcolo della media armonica.
}
\usage{
mean.a(x, \dots)
}
\arguments{
  \item{x}{vettore di dati}
  \item{\dots}{parametri aggiuntivi (ignorati in questa versione)}  
}
\seealso{
  \code{\link{mean.g}}.
}
\examples{
x <- c(1,3,2,4,6,7)
mean.a(x)
}
\keyword{univar}
