\name{mean.g}
\alias{mean.g}
\title{Calcola la media geometrica}
\description{
  Questa funzione permette il calcolo della media geometrica.
}
\usage{
mean.g(x,\dots)
}
\arguments{
  \item{x}{vettore di dati}
  \item{\dots}{parametri aggiuntivi (ignorati in questa versione)}  
}
\seealso{
  \code{\link{mean.a}}.
}
\examples{
x <- c(1,3,2,4,6,7)
mean.g(x)
}
\keyword{univar}
