\name{meana}
\alias{meana}
\title{Calcola la media armonica}
\description{
  Questa funzione permette il calcolo della media armonica.
}
\usage{
meana(x, \dots)
}
\arguments{
  \item{x}{vettore di dati}
  \item{\dots}{parametri aggiuntivi (ignorati in questa versione)}  
}
\seealso{
  \code{\link{meang}}.
}
\examples{
x <- c(1,3,2,4,6,7)
meana(x)
}
\keyword{univar}
