\name{meang}
\alias{meang}
\title{Calcola la media geometrica}
\description{
  Questa funzione permette il calcolo della media geometrica.
}
\usage{
meang(x,\dots)
}
\arguments{
  \item{x}{vettore di dati}
  \item{\dots}{parametri aggiuntivi (ignorati in questa versione)}  
}
\seealso{
  \code{\link{meana}}.
}
\examples{
x <- c(1,3,2,4,6,7)
meang(x)
}
\keyword{univar}
